+++ backend/codelight/src/main/java/com/project/codelight/global/exception/dto/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package com.project.codelight.global.exception;
+package com.project.codelight.global.exception.dto;
 
 public record ExceptionResponse(String httpMethod, String path, String codeLightCode,
                                 String message) {
